requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
flask>=3.0.0
python-docx>=1.0.0
PyPDF2>=3.0.0
fpdf2>=2.7.0
